Fixed transpose issue.
Simple, stupid swap x & y in one place. Now the low-level stuff is working right, so it's on to actually looking at algorithms.

My genetic algorithm seems too slow. It tends to get hung around 33% of perfect, which is not. I've made some tweaks so it continuously increases, but it takes a long time.