Added support for multiple target images.
Reading in and resizing images to have a good library has a lot of overhead, so it's a lot faster to be able to run the algorithm at different thresholds with different target images using the same source library of images. I've amended the command line options and main.cpp to make that easy.